+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 - feat(photos): Add Google Photos import via Picker API #356 @AhsanIsEpic
+- feat(contacts): Add support for continuous synchronization of contacts (as well as calendars) #5 @MarcelRobitaille
 
 ### Changed
 
@@ -17,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 
 - Fix: throw UnknownActivityException on Notifier #358 @lukasdotcom
+- Fix: don't wipe calendar on Google API error and improve error messages #51 @MarcelRobitaille
 
 ## [4.2.0] - 2026-06-17